moving last example and removing load_demo from test_all_examples
