+++ integrationtests/test_all_examples.py
@@ -283,31 +283,6 @@ def run_file(file_path):
         exec(content, globals)
 
 
-def load_demo(file_path, variable_name="demo"):
-    """ Loads a demo example from given file_path. Extracts the relevant
-    object via variable_name.
-
-    Parameters
-    ----------
-    file_path : str
-        The file_path of the file to be loaded
-    variable_name : str
-        The key in the global symbol state corresponding to the object of
-        interest for the demo.
-
-    Returns
-    -------
-    Instance of HasTraits
-        It is expected that this object will have edit_traits called on it,
-        so that the demo can be tested.
-    """
-    with open(file_path, "r", encoding="utf-8") as f:
-        content = f.read()
-    globals_ = globals().copy()
-    exec(content, globals_)
-    return globals_[variable_name]
-
-
 # =============================================================================
 # load_tests protocol for unittest discover
 # =============================================================================
@@ -383,204 +358,3 @@ class TestExample(unittest.TestCase):
                         reason=reason, file_path=file_path
                     )
                 )
-
-
-@requires_toolkit([ToolkitName.qt, ToolkitName.wx])
-class TestInteractExample(unittest.TestCase):
-    """ Test examples with more interactions."""
-
-    @requires_toolkit([ToolkitName.qt])
-    def test_button_editor_demo(self):
-        from pyface.ui.qt4.util.modal_dialog_tester import ModalDialogTester
-        from pyface.constant import OK
-
-        # Test ButtonEditor_demo.py in examples/demo/Standard_Edtiors
-        filepath = os.path.join(
-            DEMO, "Standard_Editors", "ButtonEditor_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-
-            simple_button = tester.find_by_id(ui, "simple")
-            custom_button = tester.find_by_id(ui, "custom")
-
-            # funcion object for instantiating ModalDialogTester should be a
-            # function that opens the dialog
-            # we want clicking the buttons to do that
-            def click_simple_button():
-                simple_button.perform(command.MouseClick())
-
-            def click_custom_button():
-                custom_button.perform(command.MouseClick())
-
-            mdtester_simple = ModalDialogTester(click_simple_button)
-            mdtester_simple.open_and_run(lambda x: x.click_button(OK))
-            self.assertTrue(mdtester_simple.dialog_was_opened)
-
-            mdtester_custom = ModalDialogTester(click_custom_button)
-            mdtester_custom.open_and_run(lambda x: x.click_button(OK))
-            self.assertTrue(mdtester_custom.dialog_was_opened)
-
-    def test_button_editor_simple_demo(self):
-        # Test ButtonEditor_simple_demo.py in examples/demo/Standard_Edtiors
-        filepath = os.path.join(
-            DEMO, "Standard_Editors", "ButtonEditor_simple_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            button = tester.find_by_name(ui, "my_button_trait")
-            for index in range(5):
-                button.perform(command.MouseClick())
-                self.assertEqual(demo.click_counter, index + 1)
-
-            click_counter = tester.find_by_name(ui, "click_counter")
-            displayed_count = click_counter.inspect(query.DisplayedText())
-            self.assertEqual(displayed_count, '5')
-
-            demo.click_counter = 10
-            displayed_count = click_counter.inspect(query.DisplayedText())
-            self.assertEqual(displayed_count, '10')
-
-    def test_converter(self):
-        # Test converter.py in examples/demo/Applications
-        filepath = os.path.join(
-            DEMO, "Applications", "converter.py"
-        )
-        demo = load_demo(filepath, "popup")
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            input_amount = tester.find_by_name(ui, "input_amount")
-            output_amount = tester.find_by_name(ui, "output_amount")
-            for _ in range(4):
-                input_amount.perform(command.KeyClick("Backspace"))
-            input_amount.perform(command.KeySequence("14.0"))
-            self.assertEqual(
-                output_amount.inspect(query.DisplayedText())[:4],
-                "1.16",
-            )
-            tester.find_by_id(ui, "Undo").perform(command.MouseClick())
-            self.assertEqual(
-                output_amount.inspect(query.DisplayedText()),
-                "1.0",
-            )
-
-    def test_checklist_editor_simple_demo(self):
-        # Test CheckListEditor_simple_demo.py in examples/demo/Standard_Edtiors
-        filepath = os.path.join(
-            DEMO, "Standard_Editors", "CheckListEditor_simple_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            checklist = tester.find_by_id(ui, "custom")
-            item3 = checklist.locate(locator.Index(2))
-            item3.perform(command.MouseClick())
-            self.assertEqual(demo.checklist, ["three"])
-            item3.perform(command.MouseClick())
-            self.assertEqual(demo.checklist, [])
-
-    def test_enum_editor_demo(self):
-        # Test EnumEditor_demo.py in examples/demo/Standard_Edtiors
-        filepath = os.path.join(
-            DEMO, "Standard_Editors", "EnumEditor_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            simple_enum = tester.find_by_id(ui, "simple")
-            simple_text_enum = tester.find_by_id(ui, "simple_text")
-            radio_enum = tester.find_by_id(ui, "radio")
-            list_enum = tester.find_by_id(ui, "list")
-            text = tester.find_by_id(ui, "text")
-            readonly = tester.find_by_id(ui, "readonly")
-
-            self.assertEqual(demo.name_list, 'A-495')
-            simple_enum.locate(locator.Index(1)).perform(command.MouseClick())
-            self.assertEqual(demo.name_list, 'A-498')
-
-            for _ in range(5):
-                simple_text_enum.perform(command.KeyClick("Backspace"))
-            simple_text_enum.perform(command.KeySequence("R-1226"))
-            simple_text_enum.perform(command.KeyClick("Enter"))
-            self.assertEqual(demo.name_list, 'R-1226')
-
-            radio_enum.locate(locator.Index(5)).perform(command.MouseClick())
-            self.assertEqual(demo.name_list, 'Foo')
-
-            list_enum.locate(locator.Index(3)).perform(command.MouseClick())
-            self.assertEqual(demo.name_list, 'TS-17')
-
-            for _ in range(5):
-                text.perform(command.KeyClick("Backspace"))
-            text.perform(command.KeySequence("A-498"))
-            text.perform(command.KeyClick("Enter"))
-            self.assertEqual(demo.name_list, 'A-498')
-
-            demo.name_list = 'Foo'
-
-            displayed_simple = simple_enum.inspect(query.DisplayedText())
-            disp_simple_text = simple_text_enum.inspect(query.DisplayedText())
-            selected_radio = radio_enum.inspect(query.SelectedText())
-            selected_list = list_enum.inspect(query.SelectedText())
-            displayed_text = text.inspect(query.DisplayedText())
-            displayed_readonly = readonly.inspect(query.DisplayedText())
-
-            displayed_selected = [
-                displayed_simple, disp_simple_text, selected_radio,
-                selected_list, displayed_text, displayed_readonly
-            ]
-            for text in displayed_selected:
-                self.assertEqual(text, 'Foo')
-
-    def test_instance_editor_demo(self):
-        # Test InstanceEditor_demo.py in examples/demo/Standard_Edtiors
-        filepath = os.path.join(
-            DEMO, "Standard_Editors", "InstanceEditor_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            simple = tester.find_by_id(ui, "simple")
-            custom = tester.find_by_id(ui, "custom")
-            occupation = custom.find_by_name("occupation")
-            occupation.perform(command.KeySequence("Job"))
-            occupation.perform(command.KeyClick("Enter"))
-            self.assertEqual(demo.sample_instance.occupation, "Job")
-
-            simple.perform(command.MouseClick())
-            name = simple.find_by_name("name")
-            name.perform(command.KeySequence("ABC"))
-            name.perform(command.KeyClick("Enter"))
-            self.assertEqual(demo.sample_instance.name, "ABC")
-
-            demo.sample_instance.name = "XYZ"
-            simple_displayed = name.inspect(query.DisplayedText())
-            custom_name = custom.find_by_name("name")
-            custom_displayed = custom_name.inspect(query.DisplayedText())
-            self.assertEqual(simple_displayed, "XYZ")
-            self.assertEqual(custom_displayed, "XYZ")
-
-    def test_list_editor_notebook_selection_demo(self):
-        # Test List_editor_notebook_selection_demo.py in examples/demo/Advanced
-        filepath = os.path.join(
-            DEMO, "Advanced", "List_editor_notebook_selection_demo.py"
-        )
-        demo = load_demo(filepath, "demo")
-
-        tester = UITester()
-        with tester.create_ui(demo) as ui:
-            people_list = tester.find_by_name(ui, "people")
-            person2 = people_list.locate(locator.Index(2))
-            person2.perform(command.MouseClick())
-            self.assertEqual(demo.index, 2)
-            age = person2.find_by_name("age")
-            age.perform(command.KeyClick("Backspace"))
-            age.perform(command.KeyClick("9"))
-            self.assertEqual(demo.people[2].age, 39)
